rename token && add comment
