+++ src/token/mod.rs
@@ -2,7 +2,7 @@
 pub mod block;
 pub mod func;
 pub mod list_block;
-pub mod operation;
+pub mod operator;
 pub mod paren_block;
 pub mod syntax;
 pub mod syntax_box;
